requirements.txt
torch
torchvision
pandas
PyYAML
Pillow
tqdm
numpy
